+++ docs/src/pages/[platform]/connected-components/storage/storageimage/examples/ErrorHandling.tsx
@@ -1,0 +1,12 @@
+import { StorageImage } from '@aws-amplify/ui-react-storage';
+
+export function App() {
+  return (
+    <StorageImage
+      alt="fallback cat"
+      path="guest/cat-in-basket.jpg"
+      fallbackSrc="/fallback_cat.jpg"
+      onGetUrlError={(error) => console.error(error)}
+    />
+  );
+}