+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @aws-amplify/ui-vue
 
+## 4.2.6
+
+### Patch Changes
+
+- [#5193](https://github.com/aws-amplify/amplify-ui/pull/5193) [`de2402842`](https://github.com/aws-amplify/amplify-ui/commit/de2402842c75e186e2c5515d20cb1873bbabc4a0) Thanks [@calebpollman](https://github.com/calebpollman)! - chore(aws-amplify): point amplify to 6.2.0
+
+- Updated dependencies [[`de2402842`](https://github.com/aws-amplify/amplify-ui/commit/de2402842c75e186e2c5515d20cb1873bbabc4a0)]:
+  - @aws-amplify/ui@6.0.14
+
 ## 4.2.5
 
 ### Patch Changes